Add print only once lines with multiple occurences
This first iteration is almost finished: the functions re_search and
print_results from the Search class are finished. The optional
parameters which modify the output (-c, -m and -u) are not implemented
yet.

Pending: re_search is not correct for 2nd+ input files, this should be
solved the same way as it has been solved in the 'main' branch:
    https://github.com/jordibellido/python-basic-grep/commit/7cdf375e5fdd941f8c04ef6bd64bffccdd01c6ab

Also pending: re_search function for the standard input stream.